requirements.txt
requests==2.32.4
pytest==8.4.1
python-dotenv==1.1.1
allure-pytest